+++ Source/Core/Actors/Timers/ActorTimer.cs
@@ -91,8 +91,7 @@ namespace Microsoft.Coyote.Actors.Timers
         }
 
         /// <summary>
-        /// Determines whether the specified System.Object is equal
-        /// to the current System.Object.
+        /// Determines whether the specified object is equal to the current object.
         /// </summary>
         public override bool Equals(object obj)
         {