+++ Source/Core/Testing/Interleaving/RandomStrategy.cs
@@ -21,66 +21,29 @@ namespace Microsoft.Coyote.Testing.Interleaving
         }
 
         /// <inheritdoc/>
-        internal override bool InitializeNextIteration(uint iteration)
-        {
-            // The random strategy just needs to reset the number of scheduled steps during
-            // the current iretation.
-            this.StepCount = 0;
-            return true;
-        }
-
-        /// <inheritdoc/>
-        internal override bool GetNextOperation(IEnumerable<ControlledOperation> ops, ControlledOperation current,
+        internal override bool NextOperation(IEnumerable<ControlledOperation> ops, ControlledOperation current,
             bool isYielding, out ControlledOperation next)
         {
             int idx = this.RandomValueGenerator.Next(ops.Count());
             next = ops.ElementAt(idx);
-            this.StepCount++;
             return true;
         }
 
         /// <inheritdoc/>
-        internal override bool GetNextBooleanChoice(ControlledOperation current, out bool next)
+        internal override bool NextBoolean(ControlledOperation current, out bool next)
         {
-            next = false;
-            if (this.RandomValueGenerator.Next(2) is 0)
-            {
-                next = true;
-            }
-
-            this.StepCount++;
+            next = this.RandomValueGenerator.Next(2) is 0 ? true : false;
             return true;
         }
 
         /// <inheritdoc/>
-        internal override bool GetNextIntegerChoice(ControlledOperation current, int maxValue, out int next)
+        internal override bool NextInteger(ControlledOperation current, int maxValue, out int next)
         {
             next = this.RandomValueGenerator.Next(maxValue);
-            this.StepCount++;
             return true;
-        }
-
-        /// <inheritdoc/>
-        internal override int GetStepCount() => this.StepCount;
-
-        /// <inheritdoc/>
-        internal override bool IsMaxStepsReached()
-        {
-            if (this.MaxSteps is 0)
-            {
-                return false;
-            }
-
-            return this.StepCount >= this.MaxSteps;
         }
 
         /// <inheritdoc/>
         internal override string GetDescription() => $"random[seed:{this.RandomValueGenerator.Seed}]";
-
-        /// <inheritdoc/>
-        internal override void Reset()
-        {
-            this.StepCount = 0;
-        }
     }
 }